requirements.txt
Jinja2==3.1.6
MarkupSafe==3.0.3
python-magic==0.4.27
PyYAML==6.0.3
scapy==2.6.1
yara-python==4.5.4
